requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
streamlit>=1.30.0
feedparser>=6.0
youtube-transcript-api>=0.6.0
pyyaml>=6.0
jinja2>=3.1
python-dateutil>=2.8
yt-dlp>=2024.1.0
python-docx>=1.0
streamlit-js-eval>=0.1.7
pytubefix>=10.0
# Optional: YouTube Data API fallback
# google-api-python-client>=2.0
